Changing from width-based sliding to margin-based
Sliding items by width has some advantages, but causes problems when
using text, because as the container narrows, the text squeezes in and
increases the height of the container. Using margins fixes this. Now
instead of applying an `active` class to the current slide, it applies
`left` to all slides left of the current one. I also simplified the
`gallop()` function and added a function to prevent a slide from
changing immediately after you click a particular slide.